Moved constants.py in fin_maestro_kin
